Fix unconstrained paths when using two ZmodScopeControllers
Using a non-unique name for the ZmodDcoClk caused a naming conflict when two ZmodScopeController IPs were used in the same design, leading to one of the IP being unconstrained. This is fixed by prepending the instance name to the clock name in the implementation TCL script used to generate an XDC file.